requirements.txt
django==2.2.7
psycopg2==2.8
django-markdownify==0.2.2
django-crispy-forms==1.7.0

# docker pull docker.elastic.co/elasticsearch/elasticsearch:7.4.2
# docker run -d -p 9200:9200 -p 9300:9300 -e "discovery.type=single-node" docker.elastic.co/elasticsearch/elasticsearch:7.4.2 
elasticsearch==7.1.0
coverage==4.4.2
factory-boy==2.9.2
